Update npm packages and configure babel
All npm packages (developmental dependencies) were updated, including
babel-core, babel-loader, moment and webpack. In addition, the
new babel-preset-es2015 package was installed and a babel configuration
file (.babelrc) was created.

This was initially motivated by wanting to update babel from version 5
to version 6. In the process of doing that, however, it became necessary
to also update both webpack and moment for reasons that were not
obvious.